requirements.txt
python-telegram-bot>=20.0a0
loguru
APScheduler
psutil
asyncio
httpx

